requirements.txt
torch==1.13.1
numpy==1.21.6
pandas==1.3.5
pytorch-lightning==1.9.0
rdkit-pypi==2022.09.4 
cmapPy==4.0.1
fcd==1.1